services/userAuth/requirements.txt
tinydb
grpcio
googleapis-common-protos
pika